+++ Mock-Get-Started/Mock-Get-Started-MyApp/src/main/java/com/microej/example/mock/MockUsageDemo.java
@@ -1,12 +1,13 @@
 /*
  * Java
  *
- * Copyright 2018-2019 MicroEJ Corp. All rights reserved.
+ * Copyright 2018-2020 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.mock;
 
 import com.microej.example.mock.style.StylesheetPopulator;
+import com.microej.example.mock.widget.KeyboardPage;
 import com.microej.example.mock.widget.MyPage;
 
 import ej.microui.MicroUI;
@@ -15,7 +16,6 @@ import ej.microui.event.generator.Pointer;
 import ej.mwt.Desktop;
 import ej.mwt.MWT;
 import ej.mwt.Panel;
-import ej.mwt.Widget;
 import ej.widget.container.transition.SlideTransitionContainer;
 import ej.widget.container.transition.TransitionContainer;
 
@@ -26,6 +26,8 @@ public class MockUsageDemo {
 
 	private static Panel Panel;
 	private static TransitionContainer TransitionContainer;
+	private static MyPage mainPage;
+	private static KeyboardPage keyboardPage;
 
 	// Prevents initialization.
 	private MockUsageDemo() {
@@ -55,7 +57,10 @@ public class MockUsageDemo {
 		TransitionContainer = new SlideTransitionContainer(MWT.LEFT, false);
 
 		// Show the main page.
-		TransitionContainer.show(new MyPage(), false);
+		mainPage = new MyPage();
+		TransitionContainer.show(mainPage, false);
+
+		keyboardPage = new KeyboardPage();
 
 		// Show the navigator.
 		Desktop desktop = new Desktop() {
@@ -79,21 +84,25 @@ public class MockUsageDemo {
 	}
 
 	/**
-	 * Gets the panel
+	 * Gets the panel.
 	 *
-	 * @return the panel
+	 * @return the panel.
 	 */
 	public static Panel getPanel() {
 		return Panel;
 	}
 
 	/**
-	 * Shows the page corresponding to the given class.
-	 *
-	 * @param page
-	 *            the page to show.
+	 * Shows the main page of the application.
 	 */
-	public static void show(Widget page) {
-		TransitionContainer.show(page, true);
+	public static void showMainPage() {
+		TransitionContainer.show(mainPage, false);
+	}
+
+	/**
+	 * Shows the keyboard page.
+	 */
+	public static void showKeyboardPage() {
+		TransitionContainer.show(keyboardPage, true);
 	}
 }