+++ Mock-Get-Started/Mock-Get-Started-MyApp/src/main/java/com/microej/example/mock/widget/Light.java
@@ -1,7 +1,7 @@
 /*
  * Java
  *
- * Copyright 2018-2019 MicroEJ Corp. All rights reserved. 
+ * Copyright 2018-2020 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.mock.widget;
@@ -21,7 +21,7 @@ public class Light extends Toggle implements OnStateChangeListener {
 	 *
 	 */
 	public Light() {
-		super(new ImageSwitch(), "Light");
+		super(new ImageSwitch(), "");
 		MySNI.switchLight(isChecked());
 		addOnStateChangeListener(this);
 	}