+++ Mock-Get-Started/Mock-Get-Started-MyApp/src/main/java/com/microej/example/mock/widget/MyPage.java
@@ -1,7 +1,7 @@
 /*
  * Java
  *
- * Copyright 2018-2019 MicroEJ Corp. All rights reserved. 
+ * Copyright 2018-2020 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.mock.widget;
@@ -9,7 +9,9 @@ package com.microej.example.mock.widget;
 import com.microej.example.mock.MockUsageDemo;
 
 import ej.widget.basic.Button;
+import ej.widget.basic.Label;
 import ej.widget.container.List;
+import ej.widget.container.Split;
 import ej.widget.listener.OnClickListener;
 import ej.widget.navigation.page.Page;
 
@@ -21,16 +23,24 @@ public class MyPage extends Page {
 	public MyPage() {
 		super();
 		List list = new List(false);
+
 		list.add(new Connected());
-		list.add(new Light());
+
+		Split lightSplit = new Split();
+		lightSplit.setFirst(new Label("Light"));
+		lightSplit.setLast(new Light());
+		list.add(lightSplit);
+
 		list.add(new TemperatureWidget());
+
 		list.add(new TextReceiver());
+
 		Button button = new Button("Send a text");
 		button.addOnClickListener(new OnClickListener() {
 
 			@Override
 			public void onClick() {
-				MockUsageDemo.show(new KeyboardPage());
+				MockUsageDemo.showKeyboardPage();
 			}
 		});
 		list.add(button);