+++ Mock-Get-Started/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.2 - 2020-09-14
+
+### Fixed
+
+   - Fix Mock-Get-Started README for MicroEJ SDK 5.x.
+
+## 1.2.1 - 2020-02-21
+
+### Fixed
+
+   - Update application.
+   
 ## 1.2.0 - 2019-11-22
 
 ### Changed
@@ -42,5 +54,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Initial revision.
 
 ---
-_Copyright 2017-2019 MicroEJ Corp. All rights reserved. 
+_Copyright 2017-2020 MicroEJ Corp. All rights reserved. 
 _Use of this source code is governed by a BSD-style license that can be found with this software._