+++ Mock-Get-Started/Mock-Get-Started-MyApp/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 - 2020-02-21
+
+### Fixed
+
+   - Updated style of the Java application.
+   - Fixed thread kept alive bug.
+
 ## 0.2.0 - 2019-11-22
 
 ### Changed
@@ -24,5 +31,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Initial revision.
 
 ---
-_Copyright 2017-2019 MicroEJ Corp. All rights reserved. 
+_Copyright 2017-2020 MicroEJ Corp. All rights reserved. 
 _Use of this source code is governed by a BSD-style license that can be found with this software._